chore: separate pages/ and components/
