--- src/components/FrontPage/index.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const FrontPage = () => {
-  return <div>Front Page</div>;
-};
-
-export default FrontPage;
