+++ src/components/Navigation/index.jsx
@@ -5,7 +5,7 @@ import * as routes from '../../constants/routes';
 
 const Navigation = () => (
   <nav>
-    <Link to='/'>Home</Link>
+    <Link to={routes.HOME}>Home</Link>
     <Link to={routes.SELECTION}>Start Planning</Link>
   </nav>
 );
