+++ src/pages/selection/candidates/index.jsx
@@ -1,13 +1,13 @@
 import React from 'react';
-import Table from '../Table';
-import dummydata from '../../constants/dummyData/candidates.json';
+import Table from 'components/Table';
+import dummydata from 'constants/dummyData/candidates.json';
 
 const Candidates = () => {
   return (
-    <div>
+    <>
       Candidates
       <Table head={dummydata.head} body={dummydata.body} />
-    </div>
+    </>
   );
 };
