+++ src/pages/selection/capital/index.jsx
