+++ src/pages/selection/index.jsx
@@ -1,24 +1,24 @@
 import React from 'react';
 import { Route } from 'react-router-dom';
 
-import SelectionNavitation from '../SelectionNavigation';
-import Candidates from '../Candidates';
-import Capital from '../Capital';
-import Schedule from '../Schedule';
-import Review from '../Review';
+import * as routes from 'constants/routes';
 
-import * as routes from '../../constants/routes';
+import Navigation from './nav';
+import Candidates from './candidates';
+import Capital from './capital';
+import Schedule from './schedule';
+import Review from './review';
 
 const SelectionPage = () => {
   return (
-    <div>
+    <>
       SelectionPage
-      <SelectionNavitation />
+      <Navigation />
       <Route path={routes.CANDIDATES} component={() => <Candidates />} />
       <Route path={routes.CAPITAL} component={() => <Capital />} />
       <Route path={routes.SCHEDULE} component={() => <Schedule />} />
       <Route path={routes.REVIEW} component={() => <Review />} />
-    </div>
+    </>
   );
 };
