+++ src/pages/selection/nav/index.jsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import * as routes from '../../constants/routes';
-
+import * as routes from 'constants/routes';
 
 const SelectionNavitation = () => {
   return (
